apps/ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery[redis]==5.4.0
redis==5.0.4
openai==1.30.0
httpx==0.27.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
structlog==24.2.0

# STT + Diarization (self-hosted path)
# openai-whisper==20231117        # uncomment for self-hosted Whisper
# pyannote.audio==3.1.1           # uncomment for speaker diarization

# NLP / ML
numpy==1.26.4
scikit-learn==1.4.2
